backend/pyproject.toml
[project]
name = "patient-safety-backend"
version = "0.1.0"
description = "GraphRAG backend for the patient safety intelligence assistant."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "google-cloud-storage>=2.18",
    "google-cloud-aiplatform>=1.71",
    "google-cloud-secret-manager>=2.21",
    "google-genai>=1.0",
    "networkx>=3.4",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
